interpolate %(branch)s format str in bucko.conf
Prior to this change, I had to hard-code the Git branch and Koji target
name in bucko.conf. This was not flexible, because we might be building
containers using composes that spanned several product releases.

Instead, add a %(branch)s format string to the configuration, and
dynamically determine the values to use for the koji section's "scm" and
"target" based on the compose metadata.

The purpose of this change is to allow us to use a single bucko.conf
file for RHCEPH 2 and 3.