RepoCompose: manage all settings in the class
Prior to this change, we were setting and getting base_product
attributes from inside and outside the RepocCompose object.

This was leading to subtle, hard-to-catch errors. For example: we were
always writing an Extras repo in all cases, even when undefined in the
config file, because __init__.py's get_compose() would set the attribute
to "None" outside the RepoCompose object, but within the object, we were
simply checking the attribute's existence (oops).

Make this consistent by moving all RepoCompose "settings" into
constructor arguments.

Move the "is_layered" assertion inside the constructor as well, so we
always run this, rather that doing it outside the class in
get_compose().

Simplify some logic in write_yum_repo_file() now, rather than cautiously
using hasattr()/getattr().