parse COMPOSE_URL format str for dist-git messages
Prior to this change, if Jenkins triggered on a dist-git CI message, we
would use a plain COMPOSE_URL string. This meant that the Jenkins job
had to construct this COMPOSE_URL string somehow before calling bucko.

The logic to convert from "dist-git branch name" -> "compose URL" is not
trivial to implement in bash, so I've built it into bucko here. Now,
when we receive dist-git messages, I can just have Jenkins set
COMPOSE_URL to a format string and let bucko interpolate it.