fix configparser support
The "import configparser" code path was really broken.

Ordinarily this would only be a problem for Python 3 support, but as it
turns out, Python 2.7.14 on Fedora 27 includes backports.configparser
(https://pypi.org/project/configparser/). On RHEL 7 "import
configparser" still fails, but Fedora will always take this path now.

Rework the configparser import stuff to be compatible with Python 2.7.5
on RHEL 7 as well as the latest Python 2 and Python 3.