fix configparser support
The "import configparser" code path was really broken.

Ordinarily this would only be a problem for Python 3 support, but as it
turns out, Fedora has a "python2-configparser" RPM that ships
backports.configparser (https://pypi.org/project/configparser/). This is
a dependency of python2-flake8.

On a bare RHEL 7 environment "import configparser" still fails, but if
python2-flake8 (and python2-configparser) is installed, Python will take
the other code path.

Rework the configparser import stuff to be compatible with Python 2.7.5
on RHEL 7 as well as the latest Python 2 and Python 3.