interpolate "distro" and "distro_upper" in COMPOSE_URL
RCM stores composes in "rhel-7" or "rhel-8" directories in NFS.  This
means that must interpolate the lower-case "rhel-8" string as well as
the upper-case "RHEL-8" string when we determine our COMPOSE_URL path.

(The change of "distro" to lower-case technically breaks bucko's current
behavior. However, with RCM's change to the new directory layout for
composes, this overall functionality was was already broken.)