make registry_host setting optional
Prior to this change, if bucko read a configuration file that did not
have a "registry_host" setting, it would fail with an error:

Problem parsing .bucko.conf: No option 'registry_host' in section: 'publish'

This setting should not be mandatory. Gracefully handle the case where
it is not present in the configuration file.