container_publisher: set REGISTRY_AUTH_FILE for podman
podman 1.6+ in RHEL 8 writes its credentials to file to
/run/user/0/containers/auth.json by default.

skopeo only reads from /run/containers/0/auth.json instead. When that
file is missing, skopeo errors out:

  Error writing manifest: Error uploading manifest latest to
  docker-registry.example.com/ceph/ceph-4.0-rhel-8: unauthorized:
  authentication required

To work around this problem, tell podman to write to the location skopeo
expects to read.

When we get more clarity in
https://bugzilla.redhat.com/show_bug.cgi?id=1800815 about the expected
behavior, maybe we can revert this.