koji_builder: log in with activate_session()
The Koji developers have deprecated the krb_login() method, and they
will remove it in Koji 1.22. Replace this with activate_session() so
that we are compatible with all authentication mechanisms.

We do not load full options from a Koji profile yet, so we must update
the opts assignment in the constructor to fill in two missing values
that activate_session() requires, "authtype" and "cert". We can simplify
this once we switch over to Koji profiles.

Update the test suite to pass "opts" through to our FakeClientSession.
Implement fake gssapi_login() and getAPIVersion() calls to satisfy
activate_session().