container_publisher: treat "skopeo copy" failures as non-fatal
Prior to this change, if "skopeo copy" failed, we would retry a few
times, and if it still failed, we would bail out without writing the
JSON record file.

The skopeo copy step is failing frequently enough that it's time to
treat this error as non-fatal.

Remove the backoff code and only try "skopeo copy" once. If it fails,
log a warning message, rather than ending the entire operation. This
allows us to continue on with publishing the JSON record of the image
for QE to test.