remove listMethods call to search for buildContainer
It is faster and less load on the Koji hub if we do not call listMethods
before buildContainer, particularly as an authenticated user.

If a user is running bucko against a Koji instance, they likely know
what they are doing, and we don't need to use Look Before You Leap here.

If the buildContainer RPC does not exist, Koji will simply return
"GenericError: Invalid method: buildContainer". If we cared about making
this error condition more user-friendly, we could catch and handle that
error, but it does not seem worth it.