always write "repository" JSON key with unique tag
Prior to this change, non-scratch builds would not result in a
"repository" key in the JSON record. Bucko would only write this key if
there was a single repo from OSBS. Non-scratch builds have more than
one repo (tag).

Update Bucko to always write a "repository" key in the JSON data,
regardless of the number of repositories OSBS presents.

Always write the unique repo (tag) into this key, so that we do not
emphasize OSBS' floating tags for QE.