search koji for list of arches for odcs
Prior to this change, we hard-coded a list of architectures in
odcs_manager. When we added aarch64 to the product, Bucko did not
generate an aarch64 repository in ODCS, so the container builds failed
for aarch64.

Remove the hard-coded list of arches in odcs_manager. Look up the list
of arches on the Koji target's build tag instead, and pass that into
odcs_manager.generate(). This matches how OSBS itself generates ODCS
composes.